% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{ManuallyEditedData_example}
\alias{ManuallyEditedData_example}
\title{retreive example type quantalife edited file name}
\usage{
ManuallyEditedData_example()
}
\value{
filename
}
\description{
retreive example type quantalife edited file name
}
