% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{mean_ddpcr}
\alias{mean_ddpcr}
\title{Take in processed csv file and calculate mean gene copy number per site per visit}
\usage{
mean_ddpcr(inputdata, n = 6, negID = "G", output = NA)
}
\arguments{
\item{inputdata}{tibble, ddpcr data processed and formatted with this package}

\item{n}{numeric, number of samples per site, default 6}

\item{negID}{character, partial string that identifies samples as negative control, these will be removed in this case, default "G"}

\item{output}{character, the name for the outputted file with mean values calculated}
}
\value{
tibble
}
\description{
Take in processed csv file and calculate mean gene copy number per site per visit
}
