% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io.R
\name{read_ddpcr}
\alias{read_ddpcr}
\title{read manually edited ddpcr data file, process it, write as a new file}
\usage{
read_ddpcr(
  filename = ManuallyEditedData_example(),
  metadata = metadata_example(),
  target = NA,
  calltype = c("auto", "manual")[2],
  output = NA
)
}
\arguments{
\item{filename}{character, the name of the file}

\item{metadata}{character, the name of the sample metadata file}

\item{target}{character, the name of the target of interest if only want to work on one}

\item{calltype}{character, if "auto", indicates the data was not manipulated in QuantaLife program. This is the default setting. If "manual", indicates some samples were manually called and csv exported}

\item{output}{character, the name for the outputted QAQC file}
}
\value{
tibble
}
\description{
read manually edited ddpcr data file, process it, write as a new file
}
